requirements.txt
blinker
celery
click
dnspython
Flask
Flask-Login
Flask-Mail
flask-marshmallow
flask-mongoengine
itsdangerous
Jinja2
MarkupSafe
marshmallow
mongoengine
pymongo
pyotp
python-dotenv
pytz
PyYAML
Unidecode
Werkzeug
WTForms
